+++ android/src/main/java/com/instabug/reactlibrary/RNInstabugReactnativeModule.java
@@ -98,7 +98,7 @@ public class RNInstabugReactnativeModule extends ReactContextBaseJavaModule {
     private final String INVOCATION_EVENT_NONE = "invocationEventNone";
     private final String INVOCATION_EVENT_SHAKE = "invocationEventShake";
     private final String INVOCATION_EVENT_SCREENSHOT = "invocationEventScreenshot";
-    private final String INVOCATION_EVENT_TWO_FINGERS_SWIPE = "invocationEventTwoFingersSwipe";
+    private final String INVOCATION_EVENT_TWO_FINGERS_SWIPE_LEFT = "invocationEventTwoFingersSwipeLeft";
     private final String INVOCATION_EVENT_FLOATING_BUTTON = "invocationEventFloatingButton";
     //InvocationModes
     private final String INVOCATION_MODE_NA = "na";
@@ -183,6 +183,10 @@ public class RNInstabugReactnativeModule extends ReactContextBaseJavaModule {
     private final String BUG_REPORTING_REPORT_TYPE_BUG = "bugReportingReportTypeBug";
     private final String BUG_REPORTING_REPORT_TYPE_FEEDBACK = "bugReportingReportTypeFeedback";
     private final String BUG_REPORTING_REPORT_TYPE_QUESTION = "bugReportingReportTypeQuestion";
+
+    private final String DISMISS_TYPE_ADD_ATTACHMENT = "dismissTypeAddAttachment";
+    private final String DISMISS_TYPE_CANCEL = "dismissTypeCancel";
+    private final String DISMISS_TYPE_SUBMIT = "dismissTypeSubmit";
 
     private final String EMAIL_FIELD_HIDDEN = "emailFieldHidden";
     private final String EMAIL_FIELD_OPTIONAL = "emailFieldOptional";
@@ -776,7 +780,7 @@ public class RNInstabugReactnativeModule extends ReactContextBaseJavaModule {
                     case INVOCATION_EVENT_FLOATING_BUTTON:
                         parsedInvocationEvents.add(InstabugInvocationEvent.FLOATING_BUTTON);
                         break;
-                    case INVOCATION_EVENT_TWO_FINGERS_SWIPE:
+                    case INVOCATION_EVENT_TWO_FINGERS_SWIPE_LEFT:
                         parsedInvocationEvents.add(InstabugInvocationEvent.TWO_FINGER_SWIPE_LEFT);
                         break;
                     case INVOCATION_EVENT_SHAKE:
@@ -813,7 +817,7 @@ public class RNInstabugReactnativeModule extends ReactContextBaseJavaModule {
         try {
             if (invocationEventValue.equals(INVOCATION_EVENT_FLOATING_BUTTON)) {
                 invocationEvent = InstabugInvocationEvent.FLOATING_BUTTON;
-            } else if (invocationEventValue.equals(INVOCATION_EVENT_TWO_FINGERS_SWIPE)) {
+            } else if (invocationEventValue.equals(INVOCATION_EVENT_TWO_FINGERS_SWIPE_LEFT)) {
                 invocationEvent = InstabugInvocationEvent.TWO_FINGER_SWIPE_LEFT;
             } else if (invocationEventValue.equals(INVOCATION_EVENT_SHAKE)) {
                 invocationEvent = InstabugInvocationEvent.SHAKE;
@@ -2421,7 +2425,7 @@ public class RNInstabugReactnativeModule extends ReactContextBaseJavaModule {
         constants.put("invocationEventNone", INVOCATION_EVENT_NONE);
         constants.put("invocationEventShake", INVOCATION_EVENT_SHAKE);
         constants.put("invocationEventScreenshot", INVOCATION_EVENT_SCREENSHOT);
-        constants.put("invocationEventTwoFingersSwipe", INVOCATION_EVENT_TWO_FINGERS_SWIPE);
+        constants.put("invocationEventTwoFingersSwipeLeft", INVOCATION_EVENT_TWO_FINGERS_SWIPE_LEFT);
         constants.put("invocationEventFloatingButton", INVOCATION_EVENT_FLOATING_BUTTON);
 
         constants.put("colorThemeLight", COLOR_THEME_LIGHT);
@@ -2439,6 +2443,10 @@ public class RNInstabugReactnativeModule extends ReactContextBaseJavaModule {
         constants.put(BUG_REPORTING_REPORT_TYPE_BUG, BUG_REPORTING_REPORT_TYPE_BUG);
         constants.put(BUG_REPORTING_REPORT_TYPE_FEEDBACK, BUG_REPORTING_REPORT_TYPE_FEEDBACK);
         constants.put(BUG_REPORTING_REPORT_TYPE_QUESTION, BUG_REPORTING_REPORT_TYPE_QUESTION);
+
+        constants.put(DISMISS_TYPE_ADD_ATTACHMENT, DISMISS_TYPE_ADD_ATTACHMENT);
+        constants.put(DISMISS_TYPE_CANCEL, DISMISS_TYPE_CANCEL);
+        constants.put(DISMISS_TYPE_SUBMIT, DISMISS_TYPE_SUBMIT);
 
         constants.put("localeArabic", LOCALE_ARABIC);
         constants.put("localeAzerbaijani", LOCALE_AZERBAIJANI);