+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## master
 
-* Fixes an issue with `setRequestFilterExpression` API when Hermes is enabled
+* Adds the ability to opt out of iOS source maps auto upload through the INSTABUG_SOURCEMAPS_UPLOAD_DISABLE env variable
+* Adds dynamic entry file support through the INSTABUG_ENTRY_FILE env variable
+* Fixes an issue with setRequestFilterExpression API not working with Hermes
+* Fixes an issue with swipe invocation event not working on Android
 
 ## 10.13.0 (2022-03-17)